+++ crates/cheatnet/src/state.rs
@@ -334,6 +334,7 @@ pub struct CheatnetState {
     pub trace_data: TraceData,
     pub encountered_errors: Vec<EncounteredError>,
     pub fuzzer_args: Vec<String>,
+    pub block_hash: HashMap<u64, Felt252>,
 }
 
 impl Default for CheatnetState {
@@ -361,6 +362,7 @@ impl Default for CheatnetState {
             },
             encountered_errors: vec![],
             fuzzer_args: Vec::default(),
+            block_hash: HashMap::default(),
         }
     }
 }