+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Forge
+
+#### Added
+
+- Cheat `cheat_block_hash` to manipulate `get_block_hash_syscall` output.
+
 ## [0.38.3] - 2025-03-07
 
 ### Forge